backend/app/requirements.txt
fastapi
sqlalchemy
uvicorn[standard]

aiofiles
pydantic-settings
pyjwt
pyjwt[crypto]

requests
bcrypt